requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
importlib_metadata==8.2.0
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
litellm==1.43.3
MarkupSafe==2.1.5
multidict==6.0.5
openai==1.40.1
packaging==24.1
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pytesseract==0.3.10
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rpds-py==0.20.0
sniffio==1.3.1
tiktoken==0.7.0
tokenizers==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
zipp==3.19.2
